+++ Bindings/Python/examples/Wrapping/exampleScholz2015GeometryPath.py
@@ -76,6 +76,11 @@ path.appendObstacle(obstacle, contact_hint)
 # adding, it defines the insertion of the path.
 path.appendPathPoint(model.getBodySet().get('b1'), osim.Vec3(-0.5, 0.1, 0.))
 
+# Enable warm starts in the wrapping solver. At each time step, the
+# wrapping solver will use the solution from the previous time step as an
+# initial guess for the current time step.
+path.setUseWarmStart(True)
+
 # Initialize the system.
 state = model.initSystem()
 model.updVisualizer().updSimbodyVisualizer().setBackgroundTypeByInt(2)