+++ OpenSim/Simulation/Model/Scholz2015GeometryPath.cpp
@@ -190,6 +190,17 @@ int Scholz2015GeometryPath::getNumPathElements() const {
     return getProperty_path_elements().size();
 }
 
+void Scholz2015GeometryPath::setUseWarmStart(bool useWarmStart) {
+    if (_index.isValid()) {
+        updCableSpan().setUseWarmStart(useWarmStart);
+    }
+    set_use_warm_start(useWarmStart);
+}
+
+bool Scholz2015GeometryPath::getUseWarmStart() const {
+    return get_use_warm_start();
+}
+
 //=============================================================================
 // ABSTRACT PATH INTERFACE
 //=============================================================================
@@ -396,6 +407,7 @@ void Scholz2015GeometryPath::extendAddToSystem(
     cable.setCurveSegmentAccuracy(1e-10);
     cable.setSolverMaxIterations(50);
     cable.setAlgorithm(SimTK::CableSpanAlgorithm::Scholz2015);
+    cable.setUseWarmStart(get_use_warm_start());
     _index = cable.getIndex();
 }
 
@@ -452,6 +464,7 @@ void Scholz2015GeometryPath::extendPostScale(const SimTK::State& s,
 //=============================================================================
 void Scholz2015GeometryPath::constructProperties() {
     constructProperty_path_elements();
+    constructProperty_use_warm_start(false);
 }
 
 const Scholz2015GeometryPathObstacle* Scholz2015GeometryPath::getObstacle(